[FEAT/#59] CalendarView 폴더링 작업 수행
